Replace Snapcraft with my new badass shell script
Configuring Snapcraft is a pain in the ass and its documentation is
clear as mud. After spending a couple hours trying to figure out how
to "cross-package", I gave up. (The final straw was finding a years old,
unanswered forum post by *one of the snap developers* asking how
to do the very thing I'm trying to do.)

It turns out that the snap file format itself is really simple: it's
just a SquashFS with an extra metadata file. Enter Snaphack, my
hot new script that converts a tarball to a SquashFS and adds that
derpy metadata file. It seems to work!

Since Zig's binary tarballs are super straightforward, this should
make it easy to start creating snaps for other supported architectures.